blacklist is now ok and tested

+++ monolith/views/users.py
@@ -55,19 +55,23 @@ def get_blacklist():
         #check user exist and that is logged in
         if request.method == 'GET':
             #show the black list of the current user
-            _user = db.session.query(blacklist.c.user_id).filter(blacklist.c.user_id==current_user.id).first()
-            if _user is not None:
+            _user = db.session.query(User.email,User.firstname,User.lastname,blacklist).filter(blacklist.c.user_id == current_user.id).filter(blacklist.c.black_id == User.id)
+            
+            if _user.first() is not None:
                 #check user has at least 1 row into the blacklist table
                 return render_template('black_list.html',action="This is your blacklist",black_list=_user)
             else:
                 return render_template('black_list.html',action="Your blacklist is empty",black_list=[])
         elif request.method == 'DELETE':
             #Clear the blacklist
-            st = blacklist.delete().where(blacklist.c.user_id == current_user.id)
-            if st is not None:
+            #1. get the blacklist of user to check if it is empty
+            black_list = db.session.query(blacklist.c.user_id).filter(blacklist.c.user_id==current_user.id).first()
+            if black_list is not None:
+                st = blacklist.delete().where(blacklist.c.user_id == current_user.id)
                 db.session.execute(st)   
-                st = db.session.query(blacklist).filter(blacklist.c.user_id == current_user.id)
-                return render_template('black_list.html',action="Your blacklist is now empty",black_list=st)
+                db.session.commit()
+                black_list = db.session.query(blacklist).filter(blacklist.c.user_id == current_user.id)
+                return render_template('black_list.html',action="Your blacklist is now empty",black_list=black_list)
             else:
                 return render_template('black_list.html',action="Your blacklist is already empty",black_list=[])
     else:
@@ -77,44 +81,43 @@ def get_blacklist():
 def add_to_black_list(target):
     #route that add target to the blacklist of user.
     if current_user is not None and hasattr(current_user, 'id'):
-        #existUser = db.session.query(User.id,blacklist.c.user_id).filter(User.id==blacklist.c.user_id).filter(User.id==current_user.id).first()   # user db obj
-        #existTarget = db.session.query(User.id,blacklist.c.black_id).filter(User.id==blacklist.c.user_id).filter(User.id==target).first()
+        #current can not add himself into the blacklist
         #check that both users are registered and that 'user' is exactly current user and nobody else
         existUser = db.session.query(User).filter(User.id==current_user.id).first()
         existTarget = db.session.query(User).filter(User.id==target).first()
         #add target into the user's blacklist
         if request.method == 'POST':
             #be sure that name is not into the blacklist already
-            if existUser is not None and existTarget is not None: 
-                try:
+            if existUser is not None and existTarget is not None and current_user.id != target: 
+                inside = db.session.query(blacklist).filter(blacklist.c.user_id == current_user.id).filter(blacklist.c.black_id == target).first()
+                if inside is None: #the user is NOT already in the blacklist
                     #try to add target into blacklist
                     existUser.black_list.append(existTarget)
                     db.session.commit()
-                    #TODO: user_bl = db.session.query(User,blacklist).filter(User.id==blacklist.c.user_id).filter(User.id==current_user.id).first() is correct? //query dubbia
-                    user_bl = db.session.query(User,blacklist,User).filter(User.id==blacklist.c.user_id).filte(User.id == blacklist.c.black_id).filter(User.id==current_user.id).first()
+                    user_bl = db.session.query(User.email,User.firstname,User.lastname,blacklist).filter(blacklist.c.user_id == current_user.id).filter(blacklist.c.black_id == User.id)
+                    #user_bl = db.session.query(User,blacklist,User).filter(User.id==blacklist.c.user_id).filter(User.id == blacklist.c.black_id).filter(User.id==current_user.id).first()
                     return render_template('black_list.html',action="User "+target+" added to the black list.",black_list = user_bl)
-                except IntegrityError:
-                    #if target already into the blacklist IntegrityError is raised by sqlAlchemy
-                    bl = db.session.query(blacklist.c.user_id).filter()
+                else: #target already in the blacklist
+                    user_bl = db.session.query(User.email,User.firstname,User.lastname,blacklist).filter(blacklist.c.user_id == current_user.id).filter(blacklist.c.black_id == User.id)
                     return render_template('black_list.html',action="This user is already in your blacklist!",black_list = user_bl)
             else:
                 #User or Target not in db
                 return render_template('black_list.html',action="Please check that you select a correct user",black_list=[])    
         elif request.method == 'DELETE':
-            #IMPORTANT: to show the user row of both current and black user we can query both and send to html a list [user,target]
-
             #Delete target from blacklist
             if existUser is not None and existTarget is not None:
                 #Delete target from current user's blacklist
-                bl_target = db.session.query(blacklist).filter(_and(blacklist.c.user_id == current_user.id, blacklist.c.black_id == target)).first()
+                bl_target = db.session.query(blacklist).filter((blacklist.c.user_id == current_user.id)&(blacklist.c.black_id == target)).first()
                 if bl_target is not None:
                     #check that target is already into the black list 
-                    bl_ = db.session.query(blacklist).filter(blacklist.c.user_id == current_user.id).first()
-                    st = blacklist.delete().where((blacklist.c.user_id == current_user)&(blacklist.c.black_id == target))
+                    
+                    st = blacklist.delete().where((blacklist.c.user_id == current_user.id)&(blacklist.c.black_id == target))
                     db.session.execute(st)
+                    db.session.commit()
+                    bl_ = db.session.query(User.email,User.firstname,User.lastname,blacklist).filter(blacklist.c.user_id == current_user.id).filter(blacklist.c.black_id == User.id)
                     return render_template('black_list.html',action = "User "+target+" removed from your black list.",black_list= bl_)
                 else:
-                    bl_user = db.session.query(blacklist).filter(blacklist.c.user_id == current_user.id).first()
+                    bl_ = db.session.query(User.email,User.firstname,User.lastname,blacklist).filter(blacklist.c.user_id == current_user.id).filter(blacklist.c.black_id == User.id)
                     
                     return render_template('black_list.html', action ="This user is not in your blacklist", black_list= bl_)
             else:
